refactor(settings): decompose the management god-file by role (WS-D)
internal/settings/management/management.go was a 919-line god-file mixing the
core service, the read-model builders, every write-path apply helper, the generic
field extractors and the link/cable-test accessors. Split it by role within the
same package (no API, signature or logic change — pure move, verified by an
identical top-level declaration set and an identical sorted code-line multiset
against the original):

- management.go (136) — Store port, Service, NewService, Get, Update, ETag, sentinels
- read.go (76)        — tier consts + buildThresholdSettings/buildCardSettings
- thresholds.go (337) — threshold + HTTP-timing apply path + the two threshold types
- extract.go (63)     — generic extractBool/String/Int/DurationField
- diagnostics.go (270)— health/speedtest/iperf/FAB/display apply path
- link.go (56)        — validLinkMode + Link/CableTest accessors

Largest resulting file is 337 LOC, so the package clears the file-size threshold
and stops being a STRICT-gate blocker for WS-Z. Monolith section-divider banner
comments dropped (each file's role is now self-evident); declaration doc comments
preserved verbatim.

Advances the Architecture Completion Plan WS-D (god-file decomposition so the
size gate can go strict).
